pyproject.toml
[project]
name = "ngd-pipeline"
version = "0.1.0"
description = "Download, process and transform NGD (National Geographic Database) data for UK address matching"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "duckdb>=1.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
    "rich>=13.7.0",
    "requests>=2.31.0",
]

[project.scripts]
ukam-ngd-build = "ngd_pipeline.cli:main"
ukam-ngd-setup = "ngd_pipeline.setup_wizard:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ngd_pipeline"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[dependency-groups]
dev = [
    "ipykernel>=7.1.0",
    "pandas>=2.3.3",
    "pyarrow>=22.0.0",
]
